# Labels: use single byte as separator - small speedup