Since `seps` is a variable, `seps[0]` has to be bounds-checked every time. Replacing with a constant everywhere it is used skips this overhead.

Benchmarks for `-tags dedupelabels`:

```
pkg: github.com/prometheus/prometheus/model/labels
                                               │ before.txt  │             after.txt             │
                                               │   sec/op    │   sec/op     vs base              │
Labels_Hash/typical_labels_under_1KB-8           153.3n ± 0%   148.6n ± 1%  -3.06% (p=0.002 n=6)
Labels_Hash/bigger_labels_over_1KB-8             185.0n ± 3%   181.9n ± 0%  -1.62% (p=0.002 n=6)
Labels_Hash/extremely_large_label_value_10MB-8   683.3µ ± 0%   681.1µ ± 0%       ~ (p=0.065 n=6)
geomean                                          2.686µ        2.641µ       -1.68%
```

(`-tags stringlabels` uses a different method to compute `Hash`)